Add scene related commands and fix compile/db2 loading.

+++ src/server/game/DataStores/DB2Structure.h
@@ -199,8 +199,8 @@ struct SceneScriptEntry
     uint32 ID;
     char const* Name;
     char const* Script;
-    uint16 PrevScriptId;
-    uint16 NextScriptId;
+    uint32 PrevScriptId;
+    uint32 NextScriptId;
 };
 
 struct SceneScriptPackageEntry
